Removing prints and adding RANDOM style
Removed excess prints of the game state to standard
out. This was originally included to debug the game
logic. Also finally implemented the 'RANDOM' play
style for the computer, in which the computer will
randomly select a cell on the board to make a move.

Also refactored some code across the files.

Signed-off-by: mrandronov <mrandronov88@gmail.com>